+++ packages/brick_rest/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Add `GzipHttpClient`. When instantiating the `RestProvider`, `gzipAllRequests` defines whether requests will be gzipped by default. Once the RestProvider is instantiated, gzipping **cannot** be disabled or configured on a per-request basis. Defaults to `false`.
+
 ## 0.0.7+1
 
 * Change `_convertJson` to protected method `convertJsonFromGet` (#57)